// src/generated/schemas/pokemonDetail.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * PokéAPI
 * All the Pokémon data you'll ever need in one place, easily accessible through a modern free open-source RESTful API.

## What is this?

This is a full RESTful API linked to an extensive database detailing everything about the Pokémon main game series.

We've covered everything from Pokémon to Berry Flavors.

## Where do I start?

We have awesome [documentation](https://pokeapi.co/docs/v2) on how to use this API. It takes minutes to get started.

This API will always be publicly available and will never require any extensive setup process to consume.

Created by [**Paul Hallett**](https://github.com/phalt) and other [**PokéAPI contributors***](https://github.com/PokeAPI/pokeapi#contributing) around the world. Pokémon and Pokémon character names are trademarks of Nintendo.
    
 * OpenAPI spec version: 2.7.0
 */
import type { PokemonDetailAbilitiesItem } from './pokemonDetailAbilitiesItem';
import type { PokemonDetailCries } from './pokemonDetailCries';
import type { PokemonDetailHeldItemsItem } from './pokemonDetailHeldItemsItem';
import type { PokemonDetailMovesItem } from './pokemonDetailMovesItem';
import type { PokemonDetailPastAbilitiesItem } from './pokemonDetailPastAbilitiesItem';
import type { PokemonDetailPastTypesItem } from './pokemonDetailPastTypesItem';
import type { PokemonDetailSprites } from './pokemonDetailSprites';
import type { PokemonDetailTypesItem } from './pokemonDetailTypesItem';
import type { PokemonFormSummary } from './pokemonFormSummary';
import type { PokemonGameIndex } from './pokemonGameIndex';
import type { PokemonSpeciesSummary } from './pokemonSpeciesSummary';
import type { PokemonStat } from './pokemonStat';

export interface PokemonDetail {
  readonly id: number;
  /** @maxLength 200 */
  name: string;
  /** @nullable */
  base_experience?: number | null;
  /** @nullable */
  height?: number | null;
  is_default?: boolean;
  /** @nullable */
  order?: number | null;
  /** @nullable */
  weight?: number | null;
  readonly abilities: readonly PokemonDetailAbilitiesItem[];
  readonly past_abilities: readonly PokemonDetailPastAbilitiesItem[];
  readonly forms: readonly PokemonFormSummary[];
  readonly game_indices: readonly PokemonGameIndex[];
  readonly held_items: readonly PokemonDetailHeldItemsItem[];
  readonly location_area_encounters: string;
  readonly moves: readonly PokemonDetailMovesItem[];
  species: PokemonSpeciesSummary;
  readonly sprites: PokemonDetailSprites;
  readonly cries: PokemonDetailCries;
  readonly stats: readonly PokemonStat[];
  readonly types: readonly PokemonDetailTypesItem[];
  readonly past_types: readonly PokemonDetailPastTypesItem[];
}